+++ wordpress-bench/files/default/wordpress_bench_client.rb
@@ -15,23 +15,21 @@ class WordpressBenchClient < Cwb::Benchmark
   end
 
   def run_scenario
-    delete_old_results
     create_properties_file
     cmd = runremote? ? run_cmd_remote : run_cmd
     stdout, stderr, status = Open3.capture3(cmd)
     raise "[wordpress-bench] #{stderr}" unless status.success?
-    results = process_results
-    @cwb.submit_metric(metric_name, timestamp, results[:average_response_time])
-    @cwb.submit_metric('wordpress-bench/num_failures', timestamp, results[:num_failures])
-    @cwb.submit_metric('wordpress-bench/failure_rate', timestamp, results[:failure_rate])
+    deliver_results('wordpress-bench/s1', 'scenario1-read-only.csv')
+    deliver_results('wordpress-bench/s2', 'scenario2-read-and-search.csv')
+    deliver_results('wordpress-bench/s3', 'scenario3-write-comment.csv')
   end
 
-  def delete_old_results
-    File.delete(results_file) if File.exist?(results_file)
-  end
-
-  def results_file
-    'scenario1-read-only.csv'
+  def deliver_results(scenario, results_file)
+    results = process_results(results_file)
+    @cwb.submit_metric("#{scenario}-response_time", timestamp, results[:response_time])
+    @cwb.submit_metric("#{scenario}-num_failures", timestamp, results[:num_failures])
+    @cwb.submit_metric("#{scenario}-failure_rate", timestamp, results[:failure_rate])
+    @cwb.submit_metric("#{scenario}-throughput", timestamp, results[:throughput])
   end
 
   def create_properties_file
@@ -66,30 +64,29 @@ class WordpressBenchClient < Cwb::Benchmark
     Time.now.to_i
   end
 
-  def metric_name
-    @cwb.deep_fetch('wordpress-bench', 'metric_name')
-  end
-
   def num_repetitions
     @cwb.deep_fetch('wordpress-bench', 'num_repetitions').to_i
   end
 
-  # NOTE: Fails if count == 0
-  def process_results
-    total = 0
-    count = 0
+  # NOTE: Fails if num_requests == 0
+  # Metrics as defined in the JMeter documentation:
+  # https://jmeter.apache.org/usermanual/glossary.html
+  def process_results(results_file)
+    total_time = 0
+    num_requests = 0
     num_failures = 0
     CSV.foreach(results_file, headers: true) do |row|
       unless is_parent(row) # skip aggregated parent rows
-        total += row['elapsed'].to_i
-        count += 1
+        total_time += row['elapsed'].to_i
+        num_requests += 1
         (num_failures += 1) if (row['success'] != 'true')
       end
     end
     results = {
-      average_response_time: (total.to_f / count),
-      num_failures: num_failures,
-      failure_rate: (num_failures.to_f / total),
+      response_time: (total_time.to_f / num_requests), # in ms
+      num_failures: num_failures, # count
+      failure_rate: (num_failures.to_f / total_time), # in %
+      throughput: num_requests.to_f / (total_time.to_f / 1000) # in seconds
     }
   end