+++ wordpress-bench/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Use new notification scheme. Upon completed execution, the load-generator (i.e., `wordpress-bench-client.rb`) instance now notifies the server where Wordpress is installed that the load testing is completed using a TCP connection.
 * Remove `http://` prefix for load-generator config
 * Update testplan according to `wordpress-bench-test-plan` (see CHANGELOG up to `0.1.5` there)
+* Submit metrics per scenario
+* Add `throughput` metric
 
 # 2.0.4 (2017-03-14)